+++ packages/docs-ui/src/controllers/doc-ui.controller.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import { Disposable, ICommandService, Inject, Injector, LifecycleStages, OnLifecycle, UniverInstanceType } from '@univerjs/core';
+import { connectInjector, Disposable, ICommandService, Inject, Injector, IUniverInstanceService, LifecycleStages, OnLifecycle, UniverInstanceType } from '@univerjs/core';
 import type { IMenuItemFactory } from '@univerjs/ui';
-import { ComponentManager, ILayoutService, IMenuService } from '@univerjs/ui';
+import { BuiltInUIPart, ComponentManager, ILayoutService, IMenuService, IShortcutService, IUIPartsService } from '@univerjs/ui';
 
 import { ITextSelectionRenderManager } from '@univerjs/engine-render';
 import { COLOR_PICKER_COMPONENT, ColorPicker } from '../components/color-picker';
@@ -30,6 +30,7 @@ import { FONT_SIZE_COMPONENT, FontSize } from '../components/font-size';
 import type { IUniverDocsUIConfig } from '../basics';
 import { CoreHeaderFooterCommand, OpenHeaderFooterPanelCommand } from '../commands/commands/doc-header-footer.command';
 import { SidebarDocHeaderFooterPanelOperation } from '../commands/operations/doc-header-footer-panel.operation';
+import { DocFooter } from '../views/doc-footer';
 import {
     AlignCenterMenuItemFactory,
     AlignJustifyMenuItemFactory,
@@ -50,6 +51,8 @@ import {
     TextColorSelectorMenuItemFactory,
     UnderlineMenuItemFactory,
 } from './menu/menu';
+import { CopyMenuFactory, CutMenuFactory, DeleteMenuFactory, PasteMenuFactory } from './menu/context-menu';
+import { AlignCenterShortCut, AlignJustifyShortCut, AlignLeftShortCut, AlignRightShortCut, BoldShortCut, BulletListShortCut, ItalicShortCut, OrderListShortCut, StrikeThroughShortCut, SubscriptShortCut, SuperscriptShortCut, UnderlineShortCut } from './shortcut/toolbar.shortcut';
 
 // FIXME: LifecycleStages.Rendered must be used, otherwise the menu cannot be added to the DOM, but the sheet ui plug-in can be added in LifecycleStages.Ready
 @OnLifecycle(LifecycleStages.Rendered, DocUIController)
@@ -60,7 +63,10 @@ export class DocUIController extends Disposable {
         @Inject(ComponentManager) private readonly _componentManager: ComponentManager,
         @ICommandService private readonly _commandService: ICommandService,
         @ILayoutService private readonly _layoutService: ILayoutService,
-        @IMenuService private readonly _menuService: IMenuService
+        @IMenuService private readonly _menuService: IMenuService,
+        @IUIPartsService private readonly _uiPartsService: IUIPartsService,
+        @IUniverInstanceService private readonly _univerInstanceService: IUniverInstanceService,
+        @IShortcutService private readonly _shortcutService: IShortcutService
     ) {
         super();
 
@@ -73,6 +79,13 @@ export class DocUIController extends Disposable {
         this.disposeWithMe(componentManager.register(FONT_FAMILY_COMPONENT, FontFamily));
         this.disposeWithMe(componentManager.register(FONT_FAMILY_ITEM_COMPONENT, FontFamilyItem));
         this.disposeWithMe(componentManager.register(FONT_SIZE_COMPONENT, FontSize));
+    }
+
+    private _initUiParts() {
+        const workbook = this._univerInstanceService.getCurrentUnitForType(UniverInstanceType.UNIVER_SHEET);
+        if (this._config.layout?.docContainerConfig?.footer && !workbook) {
+            this.disposeWithMe(this._uiPartsService.registerComponent(BuiltInUIPart.FOOTER, () => connectInjector(DocFooter, this._injector)));
+        }
     }
 
     private _initMenus(): void {
@@ -103,6 +116,38 @@ export class DocUIController extends Disposable {
         ).forEach((factory) => {
             this.disposeWithMe(this._menuService.addMenuItem(this._injector.invoke(factory), menu));
         });
+
+        [
+            CopyMenuFactory,
+            CutMenuFactory,
+            PasteMenuFactory,
+            DeleteMenuFactory,
+        ].forEach((factory) => {
+            try {
+                this.disposeWithMe(this._menuService.addMenuItem(this._injector.invoke(factory), menu));
+            } catch (error) {
+
+            }
+        });
+    }
+
+    private _initShortCut() {
+        [
+            BoldShortCut,
+            ItalicShortCut,
+            UnderlineShortCut,
+            StrikeThroughShortCut,
+            SubscriptShortCut,
+            SuperscriptShortCut,
+            AlignCenterShortCut,
+            AlignJustifyShortCut,
+            AlignRightShortCut,
+            AlignLeftShortCut,
+            OrderListShortCut,
+            BulletListShortCut,
+        ].forEach((shortcut) => {
+            this.disposeWithMe(this._shortcutService.registerShortcut(shortcut));
+        });
     }
 
     private _init(): void {
@@ -110,6 +155,8 @@ export class DocUIController extends Disposable {
         this._initMenus();
         this._initFocusHandler();
         this._initCommands();
+        this._initUiParts();
+        this._initShortCut();
     }
 
     private _initCommands(): void {