+++ packages/thread-comment-ui/src/plugin.ts
@@ -20,7 +20,6 @@ import { DependentOn, ICommandService, Inject, Injector, mergeOverrideWithDepend
 import { PLUGIN_NAME } from './types/const';
 import { ThreadCommentPanelService } from './services/thread-comment-panel.service';
 import { SetActiveCommentOperation, ToggleSheetCommentPanelOperation } from './commands/operations/comment.operations';
-import { ThreadCommentUIController } from './controllers/thread-comment-ui.controller';
 import { IThreadCommentMentionDataService, ThreadCommentMentionDataService } from './services/thread-comment-mention-data.service';
 
 export interface IUniverThreadCommentUIConfig {
@@ -42,7 +41,6 @@ export class UniverThreadCommentUIPlugin extends Plugin {
 
     override onStarting(injector: Injector): void {
         (mergeOverrideWithDependencies([
-            [ThreadCommentUIController],
             [ThreadCommentPanelService],
             [IThreadCommentMentionDataService, { useClass: ThreadCommentMentionDataService }],
         ], this._config?.overrides) as Dependency[]).forEach((dep) => {